requirements.txt
tensorflow-gpu==2.1.0
Pillow